Set currentlySelectedClassIndex to -2 when there is no class
When there is no class, filteredStudents is an empty list, thus UI can
display this list straightaway.

However, this should be handled by "add" command in the way that
when it sees currentlySelectedClassIndex is -2, throw an exception
immediately, instead of trying to add any student to a non-existing
class.

"deleteClass" command should reset currentlySelectedClassIndex to -2
as well when the only class is deleted.